MAN-21 - match eventNumber data type (#4410)

+++ projects/manage-supervision-and-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/service/SentenceService.kt
@@ -62,7 +62,7 @@ class SentenceService(
     }
 
     fun Event.toSentenceSummary() = SentenceSummary(
-        eventNumber.toLong(),
+        eventNumber,
         disposal?.type?.description ?: "Pre-Sentence"
     )
